+++ mobile/native/src/ln_dlc/mod.rs
@@ -710,16 +710,16 @@ pub fn get_unused_address() -> String {
     state::get_node().inner.get_unused_address().to_string()
 }
 
-pub fn close_channel(is_force_close: bool) -> Result<()> {
+pub async fn close_channel(is_force_close: bool) -> Result<()> {
+    tracing::info!(force = is_force_close, "Offering to close a channel");
     let node = state::try_get_node().context("failed to get ln dlc node")?;
 
-    let channels = node.inner.list_channels();
+    let channels = node.inner.list_dlc_channels()?;
     let channel_details = channels.first().context("No channel to close")?;
 
     node.inner
-        .close_channel(channel_details.channel_id, is_force_close)?;
-
-    Ok(())
+        .close_dlc_channel(channel_details.channel_id, is_force_close)
+        .await
 }
 
 pub fn collaborative_revert_channel(